feat: add reusable loading spinner (#33)
