+++ reShutCLI/AutoUpdater.cs
@@ -19,7 +19,10 @@ namespace reShutCLI
                 using var client = new HttpClient();
                 client.DefaultRequestHeaders.UserAgent.ParseAdd(UserAgent);
 
-                var response = await client.GetAsync(RepositoryUrl);
+                var response = await LoadingSpinner.RunAsync(
+                    () => client.GetAsync(RepositoryUrl),
+                    "Checking for updates...",
+                    2000);
 
                 if (response.IsSuccessStatusCode)
                 {
@@ -33,7 +36,10 @@ namespace reShutCLI
                     {
                         var asset = release.assets[0]; // Assuming the first asset is the installer
                         var downloadUrl = asset.browser_download_url;
-                        var installerPath = await DownloadInstaller(downloadUrl);
+                        var installerPath = await LoadingSpinner.RunAsync(
+                            () => DownloadInstaller(downloadUrl),
+                            "Downloading update...",
+                            2000);
 
                         if (!string.IsNullOrEmpty(installerPath))
                         {
